int=double

+++ Homework2/Program.cs
@@ -11,14 +11,14 @@ namespace ConsoleApp4
         static void Main(string[] args)
         {
             Console.Write("\n Введите сколько у вас денег =  ");
-            int n = int.Parse(Console.ReadLine());
+            double  n = double.Parse(Console.ReadLine());
 
             Console.WriteLine("\n1 бутылка воды стоит 2$\n1 булочка стоит = 1$ ");
             Console.Write("\n Сколько бутылок воды желаете приобрести? = ");
-            int waterCol = int.Parse(Console.ReadLine());
+            double  waterCol = double.Parse(Console.ReadLine());
 
             Console.Write("\n Сколько булочек хотите скушать? = ");
-            int cookiesCol = int.Parse(Console.ReadLine());
+            double  cookiesCol = double.Parse(Console.ReadLine());
 
             // цена
             int water = 2 * waterCol;
